==> 02-scripts/final_scripts/01_liana_updated_v1.ipynb <==
# LIANA tumor vs normal core atlas v1

## 00. Libraries

import numpy as np
import pandas as pd
import scanpy as sc

#import plotnine as p9

#import liana as li
import decoupler as dc
#import omnipath as op

# Import DESeq2
from pydeseq2.dds import DeseqDataSet
from pydeseq2.ds import DeseqStats

adata = sc.read_h5ad(
    "/data/projects/2022/CRCA/results/v1/final/h5ads/paired_tumor-adata.h5ad"
)

sample_key = 'sample_id'
groupby = 'cell_type_fine'
condition_key = 'sample_type'

## 01. Pseudobulk

pdata = dc.get_pseudobulk(
    adata,
    sample_col=sample_key,
    groups_col=groupby,
    layer='counts',
    mode='sum',
    min_cells=10,
    min_counts=10000
)
pdata

## 02. Subset the adata to remove the cell types that do not have both tumor and normal 

# First, group the data by 'cell_type_fine' and check which cell types have both 'tumor' and 'normal'
valid_cell_types = []

# Iterate over unique cell types
for cell_type in adata.obs['cell_type_fine'].unique():
    # Subset the data for the current cell type
    subset = adata[adata.obs['cell_type_fine'] == cell_type]
    
    # Check if both 'tumor' and 'normal' are present
    if all(x in subset.obs['sample_type'].values for x in ['tumor', 'normal']):
        valid_cell_types.append(cell_type)

# Subset the AnnData object to keep only the valid cell types
adata_subset = adata[adata.obs['cell_type_fine'].isin(valid_cell_types)].copy()

## 03. PYDESEQ2

adata = adata_subset

dea_results = {}
quiet = True

for cell_group in pdata.obs[groupby].unique():
    try:
        # Select cell profiles
        ctdata = pdata[pdata.obs[groupby] == cell_group].copy()

        # Obtain genes that pass the edgeR-like thresholds
        # NOTE: QC thresholds might differ between cell types, consider applying them by cell type
        genes = dc.filter_by_expr(ctdata,
                                  group=condition_key,
                                  min_count=5, # a minimum number of counts in a number of samples
                                  min_total_count=10 # a minimum total number of reads across samples
                                  )

        # Filter by these genes
        ctdata = ctdata[:, genes].copy()

        # Build DESeq2 object
        # NOTE: this data is actually paired, so one could consider fitting the patient label as a confounder
        dds = DeseqDataSet(
            adata=ctdata,
            counts=pdata.X + 1,
            metadata=pdata.obs,
            design_factors="sample_type",
            ref_level=["sample_type", 'normal'], # set control as reference
            refit_cooks=True,
            quiet=quiet
        )

        # Compute LFCs
        dds.deseq2()
        # Contrast between stim and ctrl
        stat_res = DeseqStats(dds, contrast=["sample-type", 'tumor', 'normal'], quiet=quiet)
        stat_res.quiet = quiet
        # Compute Wald test
        stat_res.summary()
        # Shrink LFCs
        stat_res.lfc_shrink(coeff='sample-type_tumor_vs_normal') # {condition_key}_cond_vs_ref

        dea_results[cell_group] = stat_res.results_df

    except Exception as e:
        print(f"An error occurred for cell_type_fine '{cell_group}': {e}")
        continue


## 04. Concatenate results across cell types

dea_df = pd.concat(dea_results)
dea_df = dea_df.reset_index().rename(columns={'level_0': groupby,'level_1':'index'}).set_index('index')
dea_df.head()

## 05. Save Differential Expression results

dea_df.to_csv("dea.csv")

## 06. Subset the adata to those and normalize the counts

adata = adata[adata.obs[condition_key]=='tumor'].copy()

sc.pp.normalize_total(adata, target_sum=1e4)
sc.pp.log1p(adata)

## 07. Combine the DEA results with the ligand-receptor interactions

lr_res = li.multi.df_to_lr(adata,
                           dea_df=dea_df,
                           resource_name='consensus', # NOTE: uses HUMAN gene symbols!
                           expr_prop=0.1, # calculated for adata as passed - used to filter interactions
                           groupby=groupby,
                           stat_keys=['stat', 'pvalue', 'padj'],
                           use_raw=False,
                           complex_col='stat', # NOTE: we use the Wald Stat to deal with complexes
                           verbose=True,
                           return_all_lrs=False,
                           )

## 08. Visualize distribution of the interactions

lr_res = lr_res.sort_values("interaction_stat", ascending=False)
lr_res['interaction_stat'].hist(bins=50)

## 09. Top interactions Wald statistic

li.pl.tileplot(liana_res=lr_res,
               fill = 'expr',
               label='padj',
               label_fun = lambda x: '*' if x < 0.05 else np.nan,
               top_n=15,
               orderby = 'interaction_stat',
               orderby_ascending = False,
               orderby_absolute = False,
               source_title='Ligand',
               target_title='Receptor',
               )

## 10. Ligand-Receptor Plot

plot = li.pl.dotplot(liana_res=lr_res,
                     colour='interaction_stat',
                     size='ligand_pvalue',
                     inverse_size=True,
                     orderby='interaction_stat',
                     orderby_ascending=False,
                     orderby_absolute=True,
                     top_n=10,
                     size_range=(0.5, 4)
                     )

# customize plot
(
    plot
    + p9.theme_bw(base_size=14)
    # fill cmap blue to red, with 0 the middle
    + p9.scale_color_cmap('RdBu_r', midpoint=0, limits=(-10, 10))
    # rotate x
    + p9.theme(axis_text_x=p9.element_text(angle=90), figure_size=(11, 6))

)
